pyproject.toml
[tool.poetry]
name = "YOUR_PROJECT_NAME"
version = "2.0.0"
description = ""
authors = ["Kristobal Junta <junta.kristobal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
alembic = "^1.3.3"
brotli = "^1.0.7"
scrapy = "^2.0.0"
pika = "^1.1.0"
pymysql = "^0.9.3"
requests = "^2.22.0"
sqlalchemy = "^1.3.13"
w3lib = "^1.21.0"
python-dotenv = "^0.10.5"
dotenv-linter = "^0.1.5"

[tool.poetry.dev-dependencies]
scrapy-new = "^0.1.5"
black = "^19.10b0"
pylint = "^2.4.4"
pre-commit = "^2.0.1"
isort = "^4.3.21"

[tool.black]
line-length = 99
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '/(\.eggs|\.git|\.venv|\.scrapy)/'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
